feat(execution): implement YieldHandler for improved yield exception processing and coroutine state management
